Changed time interval constants to long.

+++ src/org/lockss/protocol/LcapComm.java
@@ -411,7 +411,7 @@ public class LcapComm implements LockssManager {
   // Receive thread
   private class ReceiveThread extends Thread {
     private boolean goOn = true;
-    private int sleep = Constants.MINUTE;
+    private long sleep = Constants.MINUTE;
     private Deadline timeout = Deadline.in(sleep);
 
     private ReceiveThread(String name) {
