+++ src/org/lockss/util/Constants.java
@@ -38,13 +38,13 @@ import java.util.*;
  */
 public interface Constants {
   /** The number of milliseconds in a second */
-  public static final int SECOND = 1000;
+  public static final long SECOND = 1000;
   /** The number of milliseconds in a minute */
-  public static final int MINUTE = 60 * SECOND;
+  public static final long MINUTE = 60 * SECOND;
   /** The number of milliseconds in an hour */
-  public static final int HOUR = 60 * MINUTE;
+  public static final long HOUR = 60 * MINUTE;
   /** The number of milliseconds in a day */
-  public static final int DAY = 24 * HOUR;
+  public static final long DAY = 24 * HOUR;
   /** The number of milliseconds in a week */
-  public static final int WEEK = 7 * DAY;
+  public static final long WEEK = 7 * DAY;
 }
